+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add `Loader.new/access/deleteGlobalRef()` methods to store JNI `Object` references in `Pointer` ([issue bytedeco/javacpp-presets#1141](https://github.com/bytedeco/javacpp-presets/issues/1141))
  * Make `Loader.findLibrary()` also search in "sun.boot.library.path" for jlink ([pull #565](https://github.com/bytedeco/javacpp/pull/565))
  * Add `__int8`, `__int16`, `__int32`, and `__int64` to `InfoMap` as "basic/types" to support combinations allowed by Visual Studio
  * Add "org.bytedeco.javacpp.cacheLibraries" system property to disable cache for libraries ([pull bytedeco/gradle-javacpp#21](https://github.com/bytedeco/gradle-javacpp/pull/21))